added log library
